acomodo de test deleteNote
