+++ src/components/NoteListPanel.test.tsx
@@ -20,19 +20,6 @@ describe('NoteListPanel component', () => {
         expect(section).toBeInTheDocument();
     });
 
-    it('single note is displayed', () => {
-        const testNote: Note = { title: 'compras', content: 'pan', id: '2gf3d4s5s' }
-        seeNotesMock.mockImplementation(() => of([testNote]));
-
-        let user = { uid: '', name: '', image: '', email: '' };
-        render(<NoteListPanel user={user} />);
-
-        expect(screen.getByTestId('article')).toBeInTheDocument();
-        expect(screen.getByText(testNote.title)).toBeInTheDocument();
-        expect(screen.getByText(testNote.content)).toBeInTheDocument();
-
-    });
-
     it('multiple notes are displayed', () => {
         const testNote: Note = { title: 'compras', content: 'pan', id: '1f2f3g3h4' }
         seeNotesMock.mockImplementation(() => of([testNote, testNote]));
